Day 1: Creating files
For today I have design what parts the project will need:
- UI for the user to insert the text for the day.
- Templates of how to introduce the text into the PDF
- A pre-stablish header with dinamic date and all the other parts (title, price...)

Also I have been starting to know the libraries I will be using and already started to design all the newspaper cover using the ReportLab library. A lot of work to be done but good workflow for first day.