Day 2: Paragraphs
Today I have only work on the parapraphs function, it most very helpful the AI, because I was gonna make the hole implementation by hand given that because of lack of well research the other day, I though that the paragraph was not implemented like I wanted for the pdf on the ReportLab library. Thanks to ChatGPT and documentation I figure out that it was already implemented most of what I needed, so I just used that.