Day 3: Weather and stuff
Today I've been finishing a lot of pending things from the main template, it already looks very good. I also started to think what things to include into the journal:
- A weather section that works automated, with an OpenWeatherMap API.
- A films "cartelera" or critic section for when I have seen a movie.
- A "fake add" section for when I've buyed a cool thing I want to say in the journal.

Already started to implement the weather stuff, but it's mearley visible here do to the fact that the API is private so I have to work on this part in separted files that I can not upload into the repostory right now. A lot of work ahead, but every time seems closer, of corse all the "fronted" UI for the user is needed, but that will be more of a plus if at least I manege the back part to work properly.