Day 5: Small advances in the columns system
Today just some time working from the airport on the columns system. It seems there is not native function for this, as I thought it may be, so I will need to implement this myself. The idea is to create some type of list with the position the last element finished so that in the next element inclusion it's already included in the position correspondance, and see how many rows can it have without surpraising the max height of the column, if so moves that rows to the next column.
I also change the title of the PDF genereted so that the name match the date, so it doesn't overwrite the last day journal if you haven't saved already.